Major reordering and combining of login/citation logic
View from `/citations` and `/djangoBackend` were combined into a single
view.py file. The app structure was swapped, such that now the overall
django application with manage.py is in the outer folder, and the
webapp resides in `/citations`. Paths in settings.py were adjusted
accordingly, and the `citations` app was properly registered as an app.
The citations data model was changed to include the django built-in
user model as a foreign key to properly associate the logged in user
with their citations.
Users are now actually logged in using the `login()` function.
Various unused files were deleted.
A `requirements.txt` file was added to facilitate setting up the app
using a python virtual environment. This is important for the eventual
AWS deploy.